refactor steps.transfer._File[Writer|Reader]
Too much steps were using this private api
We make a real API, and put it in the protocol

This is a first step into not leaking pb.Reference in the protocol
