+++ master/buildbot/buildslave/protocols/base.py
@@ -68,6 +68,9 @@ class Connection(object):
         raise NotImplementedError
 
 
+# RemoteCommand base implementation and base proxy
+
+
 class RemoteCommandImpl(object):
 
     def remote_update(self, updates):
@@ -88,3 +91,63 @@ class RemoteCommandProxy(object):
 
     def remote_complete(self, failure=None):
         return self.impl.remote_complete(failure)
+
+
+# FileWriter base implementation and base proxy
+
+class FileWriterImpl(object):
+
+    def remote_write(self, data):
+        raise NotImplementedError
+
+    def remote_utime(self, accessed_modified):
+        raise NotImplementedError
+
+    def remote_unpack(self):
+        raise NotImplementedError
+
+    def remote_close(self):
+        raise NotImplementedError
+
+
+class FileWriterProxy(object):
+
+    def __init__(self, impl):
+        assert isinstance(impl, FileWriterImpl)
+        self.impl = impl
+
+    def remote_write(self, data):
+        return self.impl.remote_write(data)
+
+    def remote_utime(self, accessed_modified):
+        return self.impl.remote_utime(accessed_modified)
+
+    def remote_unpack(self):
+        return self.impl.remote_unpack()
+
+    def remote_close(self):
+        return self.impl.remote_close()
+
+# FileReader base implementation and base proxy
+
+
+class FileReaderImpl(object):
+
+    def remote_read(self, maxLength):
+        raise NotImplementedError
+
+    def remote_close(self):
+        raise NotImplementedError
+
+
+class FileReaderProxy(object):
+
+    def __init__(self, impl):
+        assert isinstance(impl, FileReaderImpl)
+        self.impl = impl
+
+    def remote_read(self, maxLength):
+        return self.impl.remote_read(maxLength)
+
+    def remote_close(self):
+        return self.impl.remote_close()
